+++ src/services/Isas.PaymentService/Models/ProductPackage.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Name { get; set; } = null!;
         public PackageType Type { get; set; }
-        public int PriceVnd { get; set; }
+        public long PriceVnd { get; set; }   // DB3 — bigint: giá pack lớn / VND > ~2,1 tỷ không tràn int
         public int? InterviewCredits { get; set; }
         public int? DurationDays { get; set; }
         public bool IsActive { get; set; }
